Added pre-race rating calculation based on qualification times

+++ Rating/Rating.Common/DataModel/Player/DriverWithoutRating.cs
@@ -8,7 +8,7 @@
         public string Name { get; set; }
 
         [XmlAttribute]
-        public double Deviation { get; set; }
+        public int Deviation { get; set; }
 
         [XmlAttribute]
         public double Volatility { get; set; }
